chore: Cleanup for v.0.0.4
