+++ src/scss/base/_variables.scss
@@ -1,5 +1,4 @@
 // Color Palette
-$text-light: #eef1e6;
 $text-dark: #333333;
 $primary-color: #150050; /* Navy Blue */
 $secondary-color: #3f0071; /* Indigo */
@@ -7,6 +6,7 @@ $accent-color: #582707; /* Burnt Sienna */
 $highlight-color: #c88304; /* Cadmium Orange */
 $background-color: #f5f3f5; /* Alabaster aka whitesmoke */
 
+$text-light: #f5f3f5;
 
 $box-color: hsl(200, 50%, 50%);
 $base-color: #0072af;
@@ -26,7 +26,7 @@ body {
 // Navigation Bar
 .navbar {
     background-color: $primary-color;
-    color: $highlight-color;
+    color: $text-light;
     padding: 1rem;
 
     ul {
