+++ src/scss/main.scss
@@ -7,7 +7,7 @@
 @import 'layout/grid';
 @import 'layout/container';
 
-@import 'pages/home';
+@import 'pages/v-0-0-3';
 @import 'pages/colorplay';
 
 @import 'themes/dark';
