+++ src/scss/utils/_functions.scss
@@ -1,0 +1,23 @@
+@function contrast($base-color, $color1, $color2) {
+    $contrast1: contrast($base-color, $color1);
+    $contrast2: contrast($base-color, $color2);
+    @return if($contrast1 > $contrast2, $color1, $color2);
+  }
+  
+  @function contrast($color1, $color2) {
+    $luminance1: luminance($color1);
+    $luminance2: luminance($color2);
+    $contrastRatio: if($luminance1 > $luminance2, $luminance1 / $luminance2, $luminance2 / $luminance1);
+    $resultColor: if($contrastRatio >= 4.5, $color1, $color2);
+    @return rgb(red($resultColor), green($resultColor), blue($resultColor));
+  }
+  
+  @function luminance($color) {
+    $r: red($color) / 255;
+    $g: green($color) / 255;
+    $b: blue($color) / 255;
+    $r: if($r <= 0.03928, $r / 12.92, pow(($r + 0.055) / 1.055, 2.4));
+    $g: if($g <= 0.03928, $g / 12.92, pow(($g + 0.055) / 1.055, 2.4));
+    $b: if($b <= 0.03928, $b / 12.92, pow(($b + 0.055) / 1.055, 2.4));
+    @return 0.2126 * $r + 0.7152 * $g + 0.0722 * $b;
+  }
