+++ src/scss/vendors/_font-awesome.scss
@@ -1,2 +1,2 @@
-@import '~@fortawesome/fontawesome-free/scss/fontawesome.scss';
-@import '~@fortawesome/fontawesome-free/scss/brands.scss';
+@import 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/scss/fontawesome.scss';
+@import 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/scss/brands.scss';
