+++ views/partials/hero.ejs
@@ -1,5 +1,10 @@
 <div class="hero">
-    <h1>Should we talk?</h1>
+    <murphy>
+    <img 
+        class="shadow1"
+        src="images/curious-pug-enterprise/curious-pug-light.png" 
+        alt="one of the power elite, the pug rules by virtue of loyalty, charisma, and genuine self-esteem" /> 
+    </murphy>
     <p>I thrive in building systems which reliably work to help people solve for speed, utility, and delight.</p>
     <button class="button-primary">Get Connected</button>
     <button class="button-secondary">Learn More</button>
