+++ views/partials/v-0-0-3/head.ejs
@@ -10,7 +10,6 @@
     <link rel="manifest" href="/site.webmanifest"> 
 
     <!-- Stylesheets -->
-    <link rel='stylesheet' href='/stylesheets/style.css' />
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
+    <link rel='stylesheet' href='/css/main.css' />
     <title><%= title %></title>
     <meta name="description" content="<%= description %>">
